requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
sqlalchemy==2.0.25
